修改了代码检查时的一些代码不规范问题

+++ src/main/java/fudan/se/lab4/service/impl/RuleServiceImpl.java
@@ -81,7 +81,7 @@ public class RuleServiceImpl implements RuleService {
                     for (Map.Entry<String, Integer> entry : sortList) {
                         Drinks drink = drinkUtil.getDrinks(entry.getKey());
                         if (num > 0 && (entry.getValue() * sub) >= num) {
-                            discount += (num / sub) * drink.getPrice() * rule.getProfit();
+                            discount += (double)((int) (num / sub)) * drink.getPrice() * rule.getProfit();
                             num -= (num / sub) * sub;
                         } else if (num > 0) {
                             discount += entry.getValue() * drink.getPrice() * rule.getProfit();
